Update the target geo
Main change in philosophy is introducing active layers as building blocks
of the tracker, which is reflected from the det construction to the paths
in the geofile. Note this means there are no , but layers.
Alternating layers have different orientation. So the concept of a
seems to be now a bit redundant. Maybe we'll get rid of it - needs discussion.
Also, we go back to the SND way of having plane 1 as verical. Commits
to handle the hit class method corresponding updates will follow.