AdvTarget geometry update with detailed module structure
Main change in philosophy is introducing active layers as building blocks
of the tracker, which is reflected from the det construction to the paths
in the geofile. Note this means there are no views, but layers.
Alternating layers have different orientation. So the concept of a view
seems to be now a bit redundant. Maybe we'll get rid of it - needs discussion.
Also, we go back to the SND way of having plane 1 as vertical.